+++ RELEASE_NOTES.md
@@ -1,6 +1,11 @@
 Release notes
 =============
 
+Changes in version 1.0.7
+========================
+  - Marked the push(Filters const&, FilterState const&) method of SearchReply and SearchListenerBase
+    as deprecated and provided push methods which take the Filters argument only.
+
 Changes in version 1.0.6
 ========================
   - Got rid of category header background, as per design (Bug #1446216).